+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## x.x.x yyyy-mm-dd
+### PaymentSheet
+* [Changed] The private beta API for https://stripe.com/docs/payments/finalize-payments-on-the-server has changed:
+  * If you use `IntentConfiguration(..., confirmHandler:)`, the confirm handler now has an additional `shouldSavePaymentMethod: Bool` parameter that you should ignore.
+  * If you use `IntentConfiguration(..., confirmHandlerForServerSideConfirmation:)`, use `IntentConfiguration(..., confirmHandler:)` instead. Additionally, the confirm handler's first parameter is now an `STPPaymentMethod` object instead of a String id. Use `paymentMethod.stripeId` to get its id and send it to your server.
 
 ## 23.8.0 2023-05-08
 ### Identity